partial retrieval + use calendar-timezone in timezone in query + put
- optimization for partial retrieval (only unmarshal calendar-data
once)

- pass calendar-timezone to query and PUT (for floating dates)

overall it is becoming a bit... bloated. The RFC is quite bloated I
suppose. It mandates some things as a MUST which seem like only MAY

For instance if someone makes a Query request and specifies a
time-window which does not match the min-date-time or max-date-time,
the server MUST reject the query. Seems a bit overkill. I think I will
break with the RFC on this point.